+++ src/main/java/kitchenpos/table/application/TableGroupService.java
@@ -2,7 +2,7 @@ package kitchenpos.table.application;
 
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.common.NotFoundEntityException;
+import kitchenpos.handler.exception.NotFoundEntityException;
 import kitchenpos.table.domain.OrderTable;
 import kitchenpos.table.domain.OrderTableRepository;
 import kitchenpos.table.domain.TableGroup;
