+++ src/test/java/kitchenpos/menu/application/MenuServiceTest.java
@@ -3,7 +3,8 @@ package kitchenpos.menu.application;
 import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
-import kitchenpos.common.NotFoundEntityException;
+import kitchenpos.handler.exception.NotCreateMenuException;
+import kitchenpos.handler.exception.NotFoundEntityException;
 import kitchenpos.menu.domain.Menu;
 import kitchenpos.menu.domain.MenuGroup;
 import kitchenpos.menu.domain.MenuGroupRepository;
@@ -13,7 +14,6 @@ import kitchenpos.menu.domain.MenuRepository;
 import kitchenpos.menu.dto.CreateMenuDto;
 import kitchenpos.menu.dto.CreateMenuProductDto;
 import kitchenpos.menu.dto.MenuDto;
-import kitchenpos.menu.exception.NotCreateMenuException;
 import kitchenpos.product.domain.Product;
 import kitchenpos.product.domain.ProductRepository;
 import org.junit.jupiter.api.BeforeEach;
